Finished CRUD functionality

+++ CPW219-AspnetMVC-CRUD-Debugging/Controllers/ProductsController.cs
@@ -17,7 +17,7 @@ namespace CPW219_AspnetMVC_CRUD_Debugging.Controllers
         {
             return View(await _context.Product.ToListAsync());
         }
-
+        
         public IActionResult Create()
         {
             return View();
@@ -29,6 +29,8 @@ namespace CPW219_AspnetMVC_CRUD_Debugging.Controllers
             if (ModelState.IsValid)
             {
                 await _context.AddAsync(product);
+                _context.SaveChanges();
+
                 return RedirectToAction(nameof(Index));
             }
             return View(product);
@@ -74,8 +76,15 @@ namespace CPW219_AspnetMVC_CRUD_Debugging.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var product = await _context.Product.FindAsync(id);
-            _context.Product.Remove(product);
-            return RedirectToAction(nameof(Index));
+
+            if (product != null)
+            {
+                _context.Product.Remove(product);
+                await _context.SaveChangesAsync();
+                return RedirectToAction(nameof(Index));
+            }
+            return RedirectToAction("Index");
+            
         }
 
         private bool ProductExists(int id)
